requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.3
    # via twitchapi
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via aiohttp
click==8.1.7
    # via uvicorn
fastapi==0.110.0
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   yarl
loguru==0.7.2
    # via -r requirements.in
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
pydantic==2.6.2
    # via fastapi
pydantic-core==2.16.3
    # via pydantic
python-dateutil==2.8.2
    # via twitchapi
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.1
    # via uvicorn
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.36.3
    # via fastapi
twitchapi==4.2.0
    # via -r requirements.in
typing-extensions==4.10.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   twitchapi
uvicorn[standard]==0.27.1
    # via -r requirements.in
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
yarl==1.9.4
    # via aiohttp
